requirements.txt
alembic==1.6.5
requests==2.26.0
asgiref==3.7.2
boto3==1.28.46
botocore==1.31.46
click==8.1.7
fastapi==0.68.2
greenlet==2.0.2
h11==0.14.0
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.1
pycryptodome==3.18.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.15.0
